# src/main/resources/static/openapi.yaml
openapi: 3.0.0
...
info:
  title: milempresas
  version: '1.0'
  description: ''
servers:
  - url: 'http://localhost:8080'
paths:
...
  '/companies/removeCompany/{companyId}':
    parameters:
      - schema:
          type: string
        name: companyId
        in: path
        required: true
    get:
      summary: removeCompany
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
      operationId: removeCompany
      description: |
        Endpoint remove company
  '/favourites/removeFavourite/{favouriteId}':
    parameters:
      - schema:
          type: string
        name: favouriteId
        in: path
        required: true
    get:
      summary: removeFavourite
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
      operationId: removeFavourite
      description: |
        Endpoint remove favourite
  '/users/removeUser/{userId}':
    parameters:
      - schema:
          type: string
        name: userId
        in: path
        required: true
    get:
      summary: removeUser
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
      operationId: removeUser
      description: Endpoint remove user
  '/categories/removeCategory/{categoryId}':
    parameters:
      - schema:
          type: string
        name: categoryId
        in: path
        required: true
    get:
      summary: removeCategory
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
      operationId: removeCategory
      description: Endpoint remove category
...